+++ integrations/opencost/CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+## 0.2.203 (2026-08-16)
+
+
+### Improvements
+
+- Add title and description metadata to OpenCost allocation and cloud-cost selector fields.
+
+
+## 0.2.202 (2026-08-16)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.48.8
+
+
 ## 0.2.201 (2026-08-13)